requirements.txt
scrapy
scrapy-playwright
openai
streamlit
pandas
plotly
scikit-learn
python-dotenv
